table: move data attribute back into Column object
Previously there was concern about double-storing column
data if it was held as an attribute of Column.  However,
without data it isn't possible to fully initialize a table
with a list of column objects.  In this implementation
the private _data attribute is set to None as soon as a
parent table reference is set, thus releasing the reference.
After the reference is set col.data returns the view
of the appropriate parent table column.