Table: support Table initialization from existing Table columns
Previously using a table column from an existing Table to create
or update a new Table would hijack that column for the new Table.
New code makes a copy of the column if it already has a
parent_table.  Added tests to show this separation is working.

Added a Column.clear_data() method to make the data clearing
process explicit instead of hiding in a parent_table property.
This allowed changing parent_table to a normal attribute.