Fix table.py problem when adding duplicate column(s)
When adding columns the Table.columns attribute was being immediately
reinitialized.  This was a problem if the subsequent code to
add the new columns raised an exception.  Re-ordered operations so
setting the actual object attributes only happens at the end after
all else has succeeded.  Wrote new tests for this issue.