Add a (skipped) test for row slice assignment.
E. Bray requested the ability to assign to a table via a row with
slice assignment, e.g.

row = table[2]
row[2:5] = [2, 3, 4]
row[:] = 3

This does not currently work because np.void (which is what numpy
returns for structured array row access) does not support slice
assignment.  Test is left there as a placeholder for now.