Changes to table initialization related to column names.
This commit implements the convention that if names or dtypes are provided
they must exactly match the number of columns.  Names can thus be used to
rename columns, except in the case of a dict input.  Names cannot be used to
select or re-order columns since this introduces too many API ambiguities
(again, dict input is an exception).  Instead one could use fancy
slicing, e.g.

    # Assume arr is a structured array with cols a, b, c
    t = Table(arr, copy=False)['c', 'b', 'a']
    t2 = Table(arr, names=('x', 'y', 'z'))['x', 'z']