Make astropy.wcs linkable from C on Windows.
This is a backport of r2510 in pywcs.

On Windows, we need to define some macros to prevent importing wcsset() from the "standard" library, which conflicts with the wcsset() in wcslib.  These macros need to be defined both for building astropy.wcs itself and for any external C libraries that link to it.