In order to accommodate a variety of dark energy equations of state, FLRWCosmology is now an abstract base class (using the Python library abc module).  In order to actually perform calculations, users must make use of a subclass of FLRWCosmology that implements a get_w(z) method (which returns the dark energy EOS at a redshift z).  The remaining functionality remains within FLRWCosmology.
A number of subclasses are provided
* LambdaCDMCosmology -- this essentially replicates the behavior
  of the old FLRWCosmology class, and assumes dark energy is a
  cosmological constant.  This should be the most common in practice.
* FlatLambdaCDMCosmology -- like LambdaCDMCosmology except it
  assumes flatness.
* wCDMCosmology -- assumes a constant dark energy equation of
  state parameterized by w = const (-1 is the same as LambdaCDM).
* w0waCDMCosmology -- assumes the CPL form for the dark energy eos:
  w(z) = w0 + wa(1-a)
* w0wzCDMCosmology -- assumes the simplest linear expansion (which
  becomes quite dangerous at all but small z): w(z) = w0 + wz*z

In most cases it will also be a good idea to overload _efunc and
_inv_efunc because the computation of things like the luminosity
distance is quite slow otherwise -- all of the above classes do
so in addition to overloading get_w.

The 2 provided default cosmologies (WMAP5/7) are now returned as
LambdaCDMCosmology objects, and the documentation has been
updated.

In addition to the above, in order to avoid the cosmological
parameters from entering into an inconsistent state, the
@property and @x.setter properties are now used to manage Om, Ok,
Ode, and H0.  So, for example, if Om is changed (e.g., cos.Om =
0.4) the value of Ok will be properly updated.  These can be
overloaded to enforce, for example, flatness (as in
FlatLambdaCDMCosmology).  And if H0 is changed .h and .critical_density0
will be updated.