Improve and test the Time object copy and replicate functionality
Previously there was no way to control whether input values were
copied or used by reference.  This affects the 'jd' format where
an input numpy array would be used by reference, thus breaking
the assumption of immutability.  Now there is a ``copy`` option
in the Time initializer.  It is still False by default, but users
who worry about this can set it to be True.

In addition the Time.copy method was previously returning an
object that shared refs with the original.  In most use cases
this is just fine since Time objects are immutable, but still
this is not really a full copy.  Now there is the ability to
make a true copy.  A new replicate() method was added to replace
the former copy() and create a replica which shares refs with
the original.  This is generally fine and memory / speed efficient.

Added tests of the above.