Refactor pprint._pformat_col into an iterator + post_processor
Split _pformat_col into an iterator section that iterates over
the string-formatted column values instead of collecting them
into a list.  The post-processing part then does stuff that is
needed for pretty-printing.

The driver for this was to use the iterator (and all the nice
formatting machinery) to do the string formatting in io.ascii.write.

Conflicts:

	astropy/table/pprint.py